Assertion fail on max_results
I forgot that parameters were passed through `calc_request_description` and converted to strings there, rather than where these methods test params (when they are still stored as-is if they aren't datetimes).

This is a slight confusion as to where in the process that parameter validation and conversion occurs. May want to refactor this down the road to make things more clear, perhaps ensure that every param in the params dict converts to strings by the time these tests would grab them.

That change would mess with some testing, but it would definitely clear things up. Plus, if we auto-convert everything to strings before passing to `calc_request_description`, A) we can look for value errors earlier and more explicitly, and B) we can remove that conversion from `calc_request_description`, making it a bit lighter.