report enum and get_reportid (#74)
* report enum and get_reportid

* removed tuples since we do now handle the decoding better

* removed get_request id

and lets see if the test now fails

* Move flake8 checking after pytest calls

A current PR is using `enum`, which was added in Py3.4. That should have produced an error for a missing module in Py2.7 testing, but `flake8` loads `enum34` as a dependency into the environment first. This contaminates the test suite prior to running tests, hiding what should be an error.

* extra require python2.7 enum34