Auto reprs for MWSDataType subclasses.
Was really starting to overthink this one.
Instead of defining a repr on every subclass that essentially does the same thing,
instead opt for an auto-repr pattern
on the abstract base class (ABC), MWSDataType.

There aren't any options here for excluding arguments with defaults,
and it's possible that something could be included from the `__dict__`
that isn't really an attribute of the class.
Particularly if there's something stored in the class
by some name other than the `__init__` arg.

I think the simplicity of this change outweighs those downsides, though.
Besides, if someone uses a design pattern in which the repr doesn't match up,
that's a style issue more than a breaking change,
something we're more likely to deal with in a code review than at runtime.