[tests] Updated fixtures for mws_credentials and api_instances
Introduces api_instance fixture, which can read the api_class
from a requesting test class in order to instantiate
the correct API for a set of tests.

In implementing this with a small change in MWS base tests,
found that mws_credentials fixture including the auth token
broke a test that required the token *not* be set.
So, broke that into a second fixture that explicitly includes
the auth_token, and added a second fixture for api_instance
that also includes the auth_token, for clarity.