Remove certain keys from these test params
Initially found a test failure due to the Timestamp not matching.
This is a minor race condition, as there are two requests being generated.

However, this isn't really a part of this particular test,
where we're more interested in the Action and other attrs
specific to the request.

It's also possible the SIgnature key could be screwy,
as that's a hash of the timestamps, as well.

So, in the end, I've thrown in most param keys
that are common to all MWS requests.
Everything left after this should be relevant to this test.