Implement InboundShipmentItem model The two models for shipment items - `InboundShipmentPlanRequestItem` and the new `InboundShipmentItem` - are similar in structure, with only minor differences in optional params and the specific name of the parameter for "quantity" ("Quantity" vs "QuantityShipped"). So, this change moves the logic originally in `InboundShipmentPlanRequestItem` to a base class, `BaseInboundShipmentItem`; which is then subclassed by both of the real data models.
Each of those classes now independently implements its optional args,
such as "asin", "condition", and "release_date"; as well as the
`param_dicts` method (required by abstract base class `MWSDataType`)

The main wrinkle was the quantity param, which has different naming
depending on which model is being used:

    "Quantity" for `InboundShipmentPlanRequestItem`
    "QuantityShipped" for `InboundShipmentItem`

This is handled by a class attr, `quantity_param`, which is injected
as a key in the param dicts within `_base_params_dict`. An assert
attempts to handle a case when someone is implementing the base
class (which they shouldn't), as we don't want to try setting
the empty string "" as a key in a dict, I think.

Further, the base class uses `_base_params_dict` instead of
implementing its own `params_dict` as an additional protection.
The abstract base class `MWSDataType` requires this method be
implemented on a subclass, otherwise it throws an error when
trying to instantiate that class. Thus, if the user attempts
to instantiate `BaseInboundShipmentItem`, they get an error; but if
they instantiate one of its subclasses, that method is defined properly,
and no error is thrown.

This should help drive user behavior to use the appropriate subclasses
instead of trying to use the base class in their own code. The only
way that can properly use it is to subclass it themselves, and at
that point we assume they know what they're doing.