Simplified from_address handling in request methods
Moved that logic out of those request methods
to a dedicated method, from_address_params
(replaces from_address_dict).

This method takes care of all the logic necessary for generating the final set of params
for the from_address, whether that comes from the API storage
or the overriding `from_address` argument.