Added docstring for a deprecation note
Not raising a warning from this method just yet (considering it),
but noting that this old interface using a setter pattern
is now just an alias for the property setter for from_address.

Going to want to remove that eventually, I think. Or just get rid of
the address storage logic entirely,
though it is rather convenient.

Anyway, this is just noting that it's not advisable to use this method.
User should just assign to from_address directly.