Permitted operations check for models.
The base datatype, MWSDataType, now includes `operations_permitted`,
a class-level list of strings that should contain the Action names
for MWS operations in which the given model may be used; and
`raise_for_operation_mismatch`, a method that checks an operation name
against that list of permitted operations. The expected convention is
to use the MWS operation's Action name, i.e. "SubmitFeed", as these
names are something we cannot alter (MWS expects those actions;
we can change the Python method names any time we want, but the
Action names should remain the same).

This is designed as an opt-in mechanism that must be implemented
in two places to take effect:

1. The model subclass must add an `operations_permitted` class with
   a list of Actions permitted for that model; and
2. The request method using that model must manually call
   `raise_for_operation_mismatch` and provide the `operation` name.

With this design, models `InboundShipmentPlanRequestItem` and
`InboundShipmentItem` implement this mechanism with definitions for
the operations permitted and the `raise_` method called within
`parse_shipment_items`. This way, if users attempt to use the
incorrect model in the CreateInboundShipmentPlan, CreateInboundShipment,
or UpdateInboundShipment operations, they will get an MWSError
to indicate the problem.

Initially I thought about adding this mechanism solely to
`BaseInboundShipmentItem`, so that those downstream models would have
that mechanism available. The more I thought about it, though,
it became clear that giving that utility to any model in the system
was a better use of that effort. Probably YAGNI for other model designs,
but the implementation was simple enough. Plus, being able to rely on
`operations_permitted` as a class attr of all datatype models *and*
having ready access to the `raise_` method without needing to write
a new one somewhere else in the module was quite helpful.

Additional changes:

- updated docstring for `parse_shipment_items` and `parse_legacy_item`
  functions.
- defined a Literal type, `ITEM_OP_TYPE`, to use in the annotations
  for those functions, hopefully to make it more clear what those
  functions are expecting.