[bug] mws_xml_to_dotdict cannot handle multiple results
In the case of a response containing multiple FooResult keys,
mws_xml_to_dotdict was trying to convert the list at the node
to a DotDict incorrectly. Now it rearranges the operations,
grabbing the results of the key first before
running `DotDict` over that result.

Test added with sample response to cover.