Fix double encoding of string parameters (#263)
Currently, the `0.8.x` branch breaks string parameters by encoding them
twice, first in the `clean_params_dict` function, and then in
`calc_request_description`. Because of this, parameters prone to have
characters that must be escaped, like `NextToken` and URLs, do not work
correctly.

https://github.com/python-amazon-mws/python-amazon-mws/pull/258 was an
attempt to fix this issue, but only for timestamp fields. Strings are
still broken.

This solution is a backport of a commit
(https://github.com/python-amazon-mws/python-amazon-mws/pull/65) that is already present in the
`develop` branch, where the `calc_request_description` function is no
longer responsible for escaping the received parameters.

It also rollbacks
https://github.com/python-amazon-mws/python-amazon-mws/pull/258,
as `clean_date` must encode the received values now.